Remove existing results endpoints. They need to be rebuilt once we get to that point
